requirements.txt
alembic==1.10.3
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.0.1
binance-connector==2.0.0
binance-futures-connector==3.3.1
black==23.3.0
blinker==1.6.2
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.2
Deprecated==1.2.13
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
flask-blueprint==1.3.0
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
install==1.3.5
itsdangerous==2.0.0
Jinja2==3.1.2
limits==3.5.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.15.1
pyOpenSSL==23.1.1
python-dotenv
pytz==2023.3
redis==4.5.5
requests==2.29.0
rich==13.3.5
service-identity==21.1.0
six==1.16.0
SQLAlchemy==2.0.10
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.0
zope.interface==6.0
